feat(BUX-188): change database pre-loading
