+++ .gitignore
@@ -1,6 +1,7 @@
 debug
 .history
 .idea/*
+.tmp
 .vscode/*
 .DS_Store
 build/
