+++ Dockerfile.releaser
@@ -7,7 +7,7 @@ RUN apt-get update && apt-get install -y wget
 WORKDIR /app
 
 ENV DB_SCHEMAPATH=./migrations
-COPY ./data/sql/migrations $DB_SCHEMAPATH
+COPY ./database/migrations $DB_SCHEMAPATH
 COPY ./entrypoint.sh .
 ARG project_name=pulse
 COPY ./$project_name .
