+++ config/config.go
@@ -139,6 +139,8 @@ type HTTP struct {
 
 // CLI represents a CLI struct.
 type CLI struct {
-	ShowVersion bool
-	ShowHelp    bool
+	ShowVersion   bool
+	ShowHelp      bool
+	ExportHeaders bool
+	ImportHeaders bool
 }
