--- config/databases/sqlite.go
@@ -1,49 +1,0 @@
-package databases
-
-import (
-	"fmt"
-	"log"
-
-	"github.com/bitcoin-sv/pulse/config"
-	"github.com/golang-migrate/migrate/v4"
-	"github.com/golang-migrate/migrate/v4/database/sqlite3"
-
-	// use blank import to use file source driver with the migrate package.
-	_ "github.com/golang-migrate/migrate/v4/source/file"
-	"github.com/jmoiron/sqlx"
-
-	// use blank import to register sqlite driver.
-	_ "github.com/mattn/go-sqlite3"
-	"github.com/pkg/errors"
-)
-
-func setupSqliteDB(c *config.Db) (*sqlx.DB, error) {
-	db, err := sqlx.Open("sqlite3", c.Dsn)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to setup database")
-	}
-	if !c.MigrateDb {
-		log.Println("migrate database set to false, skipping migration")
-		return db, nil
-	}
-	log.Println("migrating database")
-	driver, err := sqlite3.WithInstance(db.DB, &sqlite3.Config{})
-	if err != nil {
-		log.Fatalf("creating sqlite3 db driver failed %s", err)
-	}
-	m, err := migrate.NewWithDatabaseInstance(
-		fmt.Sprintf("file://%s", c.SchemaPath), "sqlite3",
-		driver)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	if err := m.Up(); err != nil {
-		if !errors.Is(err, migrate.ErrNoChange) {
-			log.Fatal(err)
-		}
-
-	}
-	log.Println("migrating database completed")
-	return db, nil
-}
