+++ dbutil/const.go
@@ -1,0 +1,6 @@
+package dbutil
+
+const (
+	compressedHeadersFilePath = "data/headers.csv.gz"
+	tmpDir                    = ".tmp"
+)
