+++ repository/webhooks_repository.go
@@ -3,7 +3,7 @@ package repository
 import (
 	"context"
 
-	"github.com/bitcoin-sv/pulse/data/sql"
+	"github.com/bitcoin-sv/pulse/database/sql"
 	"github.com/bitcoin-sv/pulse/notification"
 	"github.com/bitcoin-sv/pulse/repository/dto"
 )
